Treat exit code 1 as error (#21)
Fixes a bug in the plugin test harness. Currently exit code `1` is an
error exit code for us. Therefore, it should be treated as an error not
a success by the plugin test harness and cause tests to fail if `qlty`
exits `1`.

In the future, we may enhance `qlty` to use exit 1 to mean "failure"
(e.g. issues were found) rather than error, in which case we will need
to change this interpretation of the `qlty` exit codes at that time.