Update linux dependent plugins definitions (#33)
Linux based plugin binaries will be available on PATH, once we merge
https://github.com/qltyai/qlty/pull/197, plus, indicating a `file_name`
will be only optional for cases when the downloaded binaries doesn't
match the plugin's name. These changes update linux based plugins
definitions for the new circumstances.

Also, adding a `command_version` entry to `gitleaks` definition. It was
missing, not sure if there was a reason for.