Fix phpstan and php-codesniffer to support Windows (#159)
Related to qltyai/qlty#3276. phpstan and php-codesniffer are not
considered an executable binaries on the Windows platform and must be
executed via php.exe. Because of this, PATH won't have an effect on the
file lookup, and since the invocation directory != the tool directory,
we need to pass the full `${linter}/phpstan` type paths to the script.